backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
python-multipart
influxdb-client
paho-mqtt